+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to semantic versioning.
 - Add `IO` module with `tap` function.
 - Add `aperture`, `slice`, `reject`, and `none` functions to the `Array` module.
 - Add `reject` function to the `Record` module.
-- Add `last`, `init`, and `slice` functions to the `String` module.
+- Add `last`, `init`, `slice`, and `lookup` functions to the `String` module.
 - Add `construct` function to the `Function` module.
 - Provide the input value to the fallback function of `guard`.